fixed member list view updated countBy method to use long instead of int

+++ src/main/java/com/zfgc/dataprovider/PmConversationDataProvider.java
@@ -323,7 +323,7 @@ public class PmConversationDataProvider extends AbstractDataProvider{
 		brUserConversationDao.updateByExample(convo, ex);
 	}
 	
-	public Integer countUnread(Integer usersId) throws RuntimeException{
+	public Long countUnread(Integer usersId) throws RuntimeException{
 		BrUserConversationDbObjExample ex = brUserConversationDao.getExample();
 		ex.createCriteria().andUsersIdEqualTo(usersId).andReadFlagEqualTo(false);
